requirements.txt
io
os
re
csv
sys
mss
PIL
fpdf
time
dash
tqdm
flask
base64
openai
shutil
asyncio
datetime
csvsearch
subprocess
pytesseract
multiprocessing
pyppeteer.errors
dash.dependencies
asyncio.windows_events
dash_bootstrap_components